Fix burn-cubecl with autotune disabled
