+++ crates/burn-cubecl/src/kernel/conv/mod.rs
@@ -9,6 +9,7 @@ mod implicit_gemm;
 
 #[cfg(feature = "autotune")]
 mod tune;
+mod tune_key;
 
 pub(crate) use conv_transpose2d::*;
 pub(crate) use conv_transpose3d::*;
@@ -23,3 +24,4 @@ pub use conv_transpose2d::{ConvTranspose2dStrategy, conv_transpose2d, layout_swa
 
 #[cfg(feature = "autotune")]
 pub(crate) use tune::*;
+pub(crate) use tune_key::*;
