+++ crates/burn-cubecl/src/kernel/reduce/mod.rs
@@ -1,5 +1,7 @@
 mod base;
+#[cfg(feature = "autotune")]
 mod tune;
 
 pub use base::*;
+#[cfg(feature = "autotune")]
 pub use tune::*;
